Time: 4 ms (38.72%) | Memory: 8.4 MB (5.16%) - LeetSync

+++ 1833-find-the-highest-altitude/find-the-highest-altitude.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int largestAltitude(vector<int>& gain) {
         short int a=0,max=0;
-        for(short int i=0;i<gain.size();i++)
+        for(int i=0;i<gain.size();i++)
         {
             a+=gain[i];
             if(a>max)
